pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ComfyUI_ZMongo"
version = "0.1.0"
description = "Professional ZMongo Legal Operations suite for ComfyUI"
requires-python = ">=3.10"
dependencies = [
    "motor>=3.4",
    "pymongo>=4.6",
    "quart>=0.19",
    "werkzeug>=3.0",
    "python-dotenv>=1.0",
    "numpy",
    "torch",
    "google-genai",
    "aiohttp",
    "langchain",
    "langchain-community"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
"ComfyUI_ZMongo" = "src"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true
include = ["*"]
exclude = [
    "tests*",
    "examples*",
    "scripts*",
    "**/dist*",
    "**/build*",
    "**/.venv*",
    "__pycache__*"
]

[tool.setuptools.package-data]
"*" = ["*.json", "*.md", "web/*.js", "web/*.css"]